% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Kør CRUD-appen (kun lokalt — host 127.0.0.1)}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional options passed to [shiny::shinyApp()].}
}
\value{
A shiny app object.
}
\description{
Kør CRUD-appen (kun lokalt — host 127.0.0.1)
}
